pyproject.toml
[project]
name = "lob-chatbot-sample"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-evaluation>=1.3.0",
    "gradio>=5.20.1",
    "pyright>=1.1.400",
    "python-dotenv>=1.0.1",
    "semantic-kernel>=1.26.1",
    "psycopg>=3.2.6",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "azure-search-documents>=11.5.2",
    "azure-identity>=1.21.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[tool.pyright]
typeCheckingMode = "basic"
deprecateTypingAliases = true
reportDeprecated = "error"
